Refactor: module
